Add route to delete candidate

+++ backend/functions/routers/candidates.js
@@ -1,8 +1,9 @@
 const app = require("../express_generator")();
-const { firestore } = require('firebase-admin')
+const { firestore } = require('firebase-admin');
+const { validateFirebaseIdToken } = require("../auth");
 
-var { EVENTS_COLLECTION, CANDIDATES_COLLECTION, CANDIDATE_CODE } = require('../constants');
-const { validateCandidateCode, deleteFile } = require('../util');
+var { EVENTS_COLLECTION, CANDIDATES_COLLECTION, CANDIDATE_CODE, COMMENTS_COLLECTION } = require('../constants');
+const { validateCandidateCode, deleteFile, isAdmin } = require('../util');
 
 // TODO: for candidates routes, make sure either:
 // the current member is an admin of the event that the candidate is in.
@@ -77,42 +78,43 @@ app.post('/apply', async (req, res) => {
     profile_pic_id = "";
 
   var db = firestore();
+
+  if (!(await validateCandidateCode(candidate_code))) {
+    res.status(404).send(`Incorrect candidate_code!`);
+    return;
+  }
+
   try {
-    var is_code_valid = await validateCandidateCode(candidate_code);
+    var eventRef = await db.collection(EVENTS_COLLECTION)
+      .where(CANDIDATE_CODE, "==", candidate_code)
+      .get();
 
-    if (is_code_valid) {
-      const candidateDocRef = await db.collection(CANDIDATES_COLLECTION).add({
-        candidate_code: candidate_code,
-        email: email,
-        name: name,
-        phone_number: phone_number,
-        biography: biography,
-        application_status: 'pending',
-        resume_id: resume_id,
-        profile_pic_id: profile_pic_id,
-        comments: [],
-      }); // creates a new candidate profile
+    var event_id = eventRef.docs[0].id;
 
-      // add the candidate to the respective event
-      await db.collection(EVENTS_COLLECTION)
-        .where(CANDIDATE_CODE, "==", candidate_code)
-        .get()
-        .then((querySnapshot) => {
-          querySnapshot.forEach(function (document) {
-            document.ref.update({
-              candidates: firestore.FieldValue.arrayUnion(candidateDocRef.id)
-            });
-          });
-        });
-      res.status(200).send({
-        candidate_id: candidateDocRef.id
-      });
-      return;
+    const candidateDocRef = await db.collection(CANDIDATES_COLLECTION).add({
+      event_id: event_id,
+      candidate_code: candidate_code,
+      email: email,
+      name: name,
+      phone_number: phone_number,
+      biography: biography,
+      application_status: 'pending',
+      resume_id: resume_id,
+      profile_pic_id: profile_pic_id,
+      comments: [],
+    }); // creates a new candidate profile
 
-    } else {
-      res.status(404).send(`Incorrect candidate_code!`);
-      return;
-    }
+    // add the candidate to the respective event
+    await db.collection(EVENTS_COLLECTION).doc(event_id)
+      .update({
+        candidates: firestore.FieldValue.arrayUnion(candidateDocRef.id)
+      })
+
+    res.status(200).send({
+      candidate_id: candidateDocRef.id
+    });
+    return;
+
   } catch (e) {
     res.status(404).send(`Error adding new candidate: ${e}`);
   }
@@ -178,18 +180,46 @@ app.get('/by_event/:event_id', async (req, res) => {
  * @name DELETE/candidate/delete
  * @function
  * @param { string } candidate_id
+ * @param { string } member_id
  * @returns { string } Status 200 success if delete is successful, 404 otherwise
  * 
  */
-app.delete('/delete', async function (req, res) {
+app.delete('/delete', validateFirebaseIdToken, async function (req, res) {
+  var member_id = req.user.uid;
   var { candidate_id } = req.body;
   var db = firestore();
 
-  try {
-    var candidateRef = await db.collection(CANDIDATES_COLLECTION).doc(candidate_id);
-    var resume_id = (await candidateRef.get()).data().resume_id;
+  if (!(await isAdmin(member_id, event_id))) {
+    res.status(404).send('Non-admins cannot delete candidates!');
+    return;
+  }
 
+  try {
+    // Deleting a candidate requires multiple steps:
+    // 1. Remove all the comments related to this specific candidate
+    // 2. Remove its candidate_id from the parent event
+    // 3. Delete the candidate's resume from Cloud Storage
+    // 4. Finally, delete the Candidate entry
+    var candidateRef = db.collection(CANDIDATES_COLLECTION).doc(candidate_id);
+    var candidateData = (await candidateRef.get()).data();
+
+    // step 1
+    candidateData.comments.forEach((comment_id) => {
+      db.collection(COMMENTS_COLLECTION).doc(comment_id).delete();
+    })
+
+    // step 2
+    db.collection(EVENTS_COLLECTION)
+      .doc(candidateData.event_id)
+      .update({
+        candidates: firestore.FieldValue.arrayRemove(candidateRef.id)
+      })
+
+    // step 3
+    var resume_id = candidateData.resume_id;
     deleteFile(`resume/${resume_id}`);
+
+    // step 4
     candidateRef.delete();
     res.status(200).send("Success!");
   } catch (e) {
